Fixed bug in FixInvalidSsaTag
