+++ src/Forms/ExportPngXml.cs
@@ -1929,7 +1929,7 @@ $DROP=[DROPVALUE]" + Environment.NewLine + Environment.NewLine +
 
             text = text.Replace("<I>", "<i>");
             text = text.Replace("</I>", "</i>");
-            text = Utilities.FixInvalidItalicTags(text);
+            text = HtmlUtil.FixInvalidItalicTags(text);
 
             text = text.Replace("<B>", "<b>");
             text = text.Replace("</B>", "</b>");
