+++ src/Logic/SubtitleFormats/ScenaristClosedCaptions.cs
@@ -1728,7 +1728,7 @@ namespace Nikse.SubtitleEdit.Logic.SubtitleFormats
                 res += "</i>";
             //res = res.Replace("aã", "ã");
             //res = res.Replace("oõ", "õ");
-            return Utilities.FixInvalidItalicTags(res);
+            return HtmlUtil.FixInvalidItalicTags(res);
         }
 
         private static string GetLetter(string hexCode)
