+++ src/Test/UtilitiesTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Nikse.SubtitleEdit.Logic;
 using Nikse.SubtitleEdit.Logic.Forms;
-
+using Nikse.SubtitleEdit.Core;
 namespace Test
 {
     [TestClass]
@@ -90,7 +90,7 @@ namespace Test
         public void FixInvalidItalicTags2()
         {
             string s1 = "Gledaj prema kameri i rici <i>zdravo!";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, "Gledaj prema kameri i rici zdravo!");
         }
 
@@ -99,7 +99,7 @@ namespace Test
         public void FixInvalidItalicTags3()
         {
             string s1 = "<i>Line 1.</i>" + Environment.NewLine + "<i>Line 2.";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, "<i>Line 1." + Environment.NewLine + "Line 2.</i>");
         }
 
@@ -108,7 +108,7 @@ namespace Test
         public void FixInvalidItalicTags4()
         {
             string s1 = "It <i>is</i> a telegram," + Environment.NewLine + "it <i>is</i> ordering an advance,";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, s1);
         }
 
@@ -117,7 +117,7 @@ namespace Test
         public void FixInvalidItalicTags5()
         {
             string s1 = "- <i>It is a telegram?</i>" + Environment.NewLine + "<i>- It is.</i>";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, "<i>- It is a telegram?" + Environment.NewLine + "- It is.</i>");
         }
 
@@ -126,7 +126,7 @@ namespace Test
         public void FixInvalidItalicTags6()
         {
             string s1 = "- <i>Text1!</i>" + Environment.NewLine + "- <i>Text2.</i>";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, "<i>- Text1!" + Environment.NewLine + "- Text2.</i>");
         }
 
@@ -135,7 +135,7 @@ namespace Test
         public void FixInvalidItalicTags7()
         {
             string s1 = "<i>- You think they're they gone?<i>" + Environment.NewLine + "<i>- That can't be.</i>";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, "<i>- You think they're they gone?" + Environment.NewLine + "- That can't be.</i>");
         }
 
@@ -144,7 +144,7 @@ namespace Test
         public void FixInvalidItalicTags8()
         {
             string s1 = "<i>- You think they're they gone?</i>" + Environment.NewLine + "<i>- That can't be.<i>";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, "<i>- You think they're they gone?" + Environment.NewLine + "- That can't be.</i>");
         }
 
@@ -153,8 +153,16 @@ namespace Test
         public void FixInvalidItalicTags9()
         {
             string s1 = "FALCONE:<i> I didn't think</i>\r\n<i>it was going to be you,</i>";
-            string s2 = Utilities.FixInvalidItalicTags(s1);
+            string s2 = HtmlUtil.FixInvalidItalicTags(s1);
             Assert.AreEqual(s2, "FALCONE: <i>I didn't think\r\nit was going to be you,</i>");
+        }
+
+        [TestMethod]
+        [DeploymentItem("SubtitleEdit.exe")]
+        public void FixUnneededSpacesDoubleSpaceNarrator1()
+        {
+            var result = Utilities.RemoveUnneededSpaces("Kurt: ... true but bad.", "en");
+            Assert.AreEqual(result, "Kurt: ...true but bad.");
         }
 
         [TestMethod]
@@ -391,5 +399,26 @@ Foobar.</font>";
             Assert.AreEqual(result, "<i>Foobar</i>");
         }
 
+        [TestMethod]
+        [DeploymentItem("SubtitleEdit.exe")]
+        public void FixInvalidSsaTagsTest1()
+        {
+            var current = "Who is {\\an7}next?";
+            var expected = "{\\an7}Who is next?";
+
+            string result = Utilities.FixInvalidSsaTags(current);
+            Assert.AreEqual(expected, result);
+        }
+
+        [TestMethod]
+        [DeploymentItem("SubtitleEdit.exe")]
+        public void RemoveSsaTagsTest1()
+        {
+            var current = "Who is {\\an7}next?";
+            var expected = "Who is next?";
+
+            string result = Utilities.RemoveSsaTags(current);
+            Assert.AreEqual(expected, result);
+        }
     }
 }
